Log volume: Move DataSourceWithLogsVolumeSupport to logs.ts
