+++ packages/grafana-data/src/types/index.ts
@@ -7,7 +7,6 @@ export * from './dashboard';
 export * from './query';
 export * from './annotations';
 export * from './logs';
-export * from './logsVolume';
 export * from './navModel';
 export * from './select';
 export * from './time';
